(hel-444) Ajout autorisation et capacites EJ
