+++ src/frontend/ui/entité-juridique/PageEntitéJuridique.tsx
@@ -3,6 +3,7 @@ import Head from "next/head";
 import { useBreadcrumb } from "../commun/hooks/useBreadcrumb";
 import { Titre } from "../commun/Titre/Titre";
 import { BlocActivitéSanitaire } from "./bloc-activité/BlocActivitéSanitaire";
+import { BlocAutorisationsCapacites } from "./bloc-autorisations-capacites/BlocAutorisationsCapacites";
 import { BlocBudgetFinance } from "./bloc-budget-finance/BlocBudgetFinance";
 import { Catégorisation } from "./catégorisation/Catégorisation";
 import { EntitéJuridiqueViewModel } from "./EntitéJuridiqueViewModel";
@@ -33,6 +34,7 @@ export const PageEntitéJuridique = ({ entitéJuridiqueViewModel, établissement
       <Titre logo={LogoEntitéJuridique}>{entitéJuridiqueViewModel.titre}</Titre>
       <BlocIdentité entitéJuridiqueViewModel={entitéJuridiqueViewModel} />
       <ListeDesÉtablissementsTerritoriauxRattachés ETRattachés={établissementsTerritoriauxRattachésViewModels} />
+      <BlocAutorisationsCapacites entitéJuridiqueAutorisationsCapacitesViewModel={entitéJuridiqueViewModel.entitéJuridiqueAutorisationsCapacitesViewModel} />
       <BlocActivitéSanitaire entitéJuridiqueActivitéViewModel={entitéJuridiqueViewModel.entitéJuridiqueActivitéViewModel} />
       <BlocBudgetFinance entitéJuridiqueBudgetFinanceViewModel={entitéJuridiqueViewModel.entitéJuridiqueBudgetFinanceViewModel} />
     </main>
